Updates to custom data and heavy updates to rvn_netcdf_to_gridshp, including fixing bug for assigning projections and transformation properly to output shapefile.

+++ R/rvn_netcdf_to_gridshp.R
@@ -5,17 +5,21 @@
 #' generates an estimate of the grid polygons associated with each netCDF cell and exports this to a shapefile (outshp)
 #'
 #' @details
-#' Note that the function can fail due to bad netCDF file or inappropriate UTM zone.
+#' projID should be provided as an integer value referring to a valid EPSG coordinate system. If projID is left NULL, the
+#' output shapefile will be left in lat/long coordinates and WGS84 projection. Note that the function can fail due to bad
+#' netCDF file or inappropriate coordinate system ID. A list of IDs for
+#' projected coordinate system may be found on the \href{https://spatialreference.org/ref/epsg/}{Spatial Reference webpage}
+#' or on the \href{http://resources.esri.com/help/9.3/arcgisserver/apis/rest/pcs.html}{ESRI webpage}.
 #'
-#' @note
-#' Sample netcdf file in RavenR package obtained from [GWF Cuizinart](https://tuna.cs.uwaterloo.ca/), metadata for the ERA5 downloaded sample data may
-#'  be found at \url{https://gwf-uwaterloo.github.io/wamedex/#13079_iso.xml}.
+#' If the outshp is NULL, the shapefile object is returned by the function and nothing is written to file. If outshp is provided,
+#' the shapefile is also written to file. The outshp should be supplied as the file name without the extension (i.e. no .shp
+#' included in the argument).
 #'
 #' @param ncfile netCDF file with latitude and longitude variables
-#' @param UTMzone UTM zone for exported shapefile (integer)
-#' @param outshp name of output shapefile prefix (i.e., no .shp extension, DEFAULT 'rvn_output_shapefile')
+#' @param projID projected coordinate system ID (EPSG numeric code) to project shapefile to (default NULL, optional)
+#' @param outshp name of output shapefile prefix (DEFAULT NULL, optional)
 #'
-#' @return \item{TRUE}{returns TRUE if executed properly. Also generates shapefile in folder}
+#' @return \item{shapefile}{returns sf shapefile object; will also write a shapefile to the outshp if provided}
 #'
 #' @seealso \code{\link{rvn_gen_gridweights}} for generating a shapefile of gridweights
 #'
@@ -23,10 +27,20 @@
 #'
 #' @examples
 #'
+#' # get location for sample netcdf file
 #' ncfile <- system.file("extdata/Nith_era5_sample.nc", package="RavenR")
-#' UTMzone <- 17
+#'
+#' # produce shapefile in lat/long
+#' myshp <- rvn_netcdf_to_gridshp(ncfile)
+#' class(myshp)
+#' sf::st_crs(myshp)$input
+#' plot(myshp$geometry)
+#'
+#' # write shapefile to file in UTM coordinates
+#' projID <- 26917 # NAD83 UTM Zone 17N, appropriate UTM zone for Nith watershed
 #' outshp <- 'Nith_output_shapefile'
-#' rvn_netcdf_to_gridshp(ncfile, UTMzone, outshp)
+#' myshp <- rvn_netcdf_to_gridshp(ncfile, projID, outshp)
+#' sf::st_crs(myshp)$input
 #'
 #'
 #' @export rvn_netcdf_to_gridshp
@@ -35,7 +49,7 @@
 #' @importFrom deldir deldir tile.list
 #' @importFrom methods slot
 #' @importFrom ncdf4 nc_open ncvar_get
-rvn_netcdf_to_gridshp <- function(ncfile,UTMzone,outshp='rvn_output_shapefile')
+rvn_netcdf_to_gridshp <- function(ncfile,projID=NULL,outshp=NULL)
 {
 
   # extract lat-long from netCDF file
@@ -54,13 +68,19 @@ rvn_netcdf_to_gridshp <- function(ncfile,UTMzone,outshp='rvn_output_shapefile')
   IDs<-(c-1)*nrows+(r-1)
   #print(c)
 
-  # transform lat-long to UTM zone
+  # build SpatialPointsDataFrame, transform lat-long to UTM zone (if warranted)
   #-----------------------------------------------------------------------
   # latlong <- data.frame(ID = 1:2, X = long, Y = lat)
-  latlong <- data.frame(ID = IDs, X = long, Y = lat)
-  sp::coordinates(latlong) <- c("X", "Y")
-  sp::proj4string(latlong) <- CRS("+proj=longlat +datum=WGS84")  ## for example
-  res <- spTransform(latlong, CRS(paste0("+proj=utm +zone=",UTMzone," ellps=WGS84")))
+  res <- data.frame(ID = IDs, X = long, Y = lat)
+  sp::coordinates(res) <- c("X", "Y")
+  sp::proj4string(res) <- sp::CRS("+proj=longlat +datum=WGS84")
+  if (!is.null(projID)) {
+    # res <- spTransform(latlong, sp::CRS(paste0("+proj=utm +zone=",UTMzone," ellps=WGS84")))
+    res <- spTransform(res, sp::CRS(sprintf("+init=epsg:%i",projID)))
+    myprojstring = sp::CRS(sprintf("+init=epsg:%i",projID))
+  } else {
+    myprojstring = sp::CRS("+proj=longlat +datum=WGS84")
+  }
   xx<-res@coords[1:length(long)]
   yy<-res@coords[length(long)+(1:length(long))]
 
@@ -85,14 +105,18 @@ rvn_netcdf_to_gridshp <- function(ncfile,UTMzone,outshp='rvn_output_shapefile')
   rnames<-sapply(slot(SP, 'polygons'), function(x) slot(x, 'ID'))
 
   voronoishp = SpatialPolygonsDataFrame(SP, data=data.frame(x=xx,y=yy, row.names=rnames,GridIDs=IDs,Latit=lat,Longit=long))
+  sp::proj4string(voronoishp) <- myprojstring # assign projection to new shapefile accordingly
 
   # write to file
   #-----------------------------------------------------------------------
-  unlink(paste0(outshp,".*")) # deletes old file if it exists
-  writeOGR(voronoishp,dsn=getwd(),layer=outshp, driver="ESRI Shapefile")
+  if (!is.null(outshp)) {
+    unlink(paste0(outshp,".*")) # deletes old file if it exists
+    writeOGR(voronoishp,dsn=getwd(),layer=outshp, driver="ESRI Shapefile")
+    print(sprintf("%s written to file", outshp))
+  }
 
-  #plot(voronoishp,col="blue")
+  voronoishp <- sf::st_as_sf(voronoishp)
 
-  return(TRUE)
+  return(voronoishp)
 }
